+++ src/common/hooks/useNumberGeneratorOptions/useNumberGeneratorOptions.js
@@ -28,13 +28,14 @@ export const useNumberGeneratorOptions = () => {
     queryKey: [namespace],
     queryFn: async () => {
       const response = await ky.get(MOD_SETTINGS_API, { searchParams }).json();
-      const settingValue = response?.items?.[0]?.value;
+      const settingValue = response?.items?.[0]?.value ?? {};
 
       return {
         accessionNumber: settingValue.accessionNumber || '',
         barcode: settingValue.barcode || '',
         callNumber: settingValue.callNumber || '',
         callNumberHoldings: settingValue.callNumberHoldings || '',
+        identifier: settingValue.identifier || '',
         useSharedNumber: settingValue.useSharedNumber ?? false,
       };
     },