+++ src/edit/identifierFields.test.js
@@ -7,6 +7,11 @@ import stripesFinalForm from '@folio/stripes/final-form';
 import renderWithRouter from '../../test/jest/helpers/renderWithRouter';
 import renderWithIntl from '../../test/jest/helpers/renderWithIntl';
 
+import {
+  NUMBER_GENERATOR_OPTIONS_OFF,
+  NUMBER_GENERATOR_OPTIONS_ON_EDITABLE,
+  NUMBER_GENERATOR_OPTIONS_ON_NOT_EDITABLE,
+} from '../settings/NumberGeneratorSettings/constants';
 import IdentifierFields from './identifierFields';
 import translationsProperties from '../../test/jest/helpers/translationsProperties';
 
@@ -18,14 +23,15 @@ const props = {
   identifierTypes: [{ name: 'identifyName', id: '129459' }]
 };
 
-const Form = ({ handleSubmit }) => (
+const Form = ({ handleSubmit, identifierFieldsProps = {} }) => (
   <form onSubmit={handleSubmit}>
-    <IdentifierFields {...props} />
+    <IdentifierFields {...props} {...identifierFieldsProps} />
   </form>
 );
 
 Form.propTypes = {
   handleSubmit: PropTypes.func.isRequired,
+  identifierFieldsProps: PropTypes.object,
 };
 
 const WrappedForm = stripesFinalForm({
@@ -34,8 +40,8 @@ const WrappedForm = stripesFinalForm({
   canDelete: true,
 })(Form);
 
-const renderIdentifierFields = () => renderWithIntl(
-  renderWithRouter(<WrappedForm onSubmit={onSubmit} />),
+const renderIdentifierFields = (identifierFieldsProps = {}) => renderWithIntl(
+  renderWithRouter(<WrappedForm onSubmit={onSubmit} identifierFieldsProps={identifierFieldsProps} />),
   translationsProperties,
 );
 
@@ -56,5 +62,45 @@ describe('IdentifierFields', () => {
     expect(myText).toHaveValue('');
     fireEvent.change(myText, { target: { value: 'Enter text' } });
     expect(myText).toHaveValue('Enter text');
+  });
+
+  describe('Number generator button', () => {
+    const addIdentifier = () => fireEvent.click(screen.getByText('Add identifier'));
+
+    describe('when number generator settings for identifier is "onNotEditable"', () => {
+      it('should render generate identifier button and disable identifier field', () => {
+        renderIdentifierFields({
+          numberGeneratorData: { identifier: NUMBER_GENERATOR_OPTIONS_ON_NOT_EDITABLE },
+        });
+        addIdentifier();
+
+        expect(screen.getByRole('button', { name: 'Generate identifier' })).toBeInTheDocument();
+        expect(screen.getByRole('textbox')).toBeDisabled();
+      });
+    });
+
+    describe('when number generator settings for identifier is "onEditable"', () => {
+      it('should render generate identifier button and enable identifier field', () => {
+        renderIdentifierFields({
+          numberGeneratorData: { identifier: NUMBER_GENERATOR_OPTIONS_ON_EDITABLE },
+        });
+        addIdentifier();
+
+        expect(screen.getByRole('button', { name: 'Generate identifier' })).toBeInTheDocument();
+        expect(screen.getByRole('textbox')).toBeEnabled();
+      });
+    });
+
+    describe('when number generator settings for identifier is "off"', () => {
+      it('should not render generate identifier button and enable identifier field', () => {
+        renderIdentifierFields({
+          numberGeneratorData: { identifier: NUMBER_GENERATOR_OPTIONS_OFF },
+        });
+        addIdentifier();
+
+        expect(screen.queryByRole('button', { name: 'Generate identifier' })).not.toBeInTheDocument();
+        expect(screen.getByRole('textbox')).toBeEnabled();
+      });
+    });
   });
 });