+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Update `Instance status term` field when opening instance after updating. Fixes UIIN-3650.
 * Show "Source" field in "Version history" for a user not having permissions in Central tenant. Fixes UIIN-3686.
 * Update Inventory results list's select/unselect all checkbox label. Refs UIIN-3684.
+* Integrate Button "Generate identifier" in Inventory instance records. Refs UIIN-3679.
 
 ## [14.0.3](https://github.com/folio-org/ui-inventory/tree/v14.0.3) (2026-05-29)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v14.0.2...v14.0.3)