+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Update `Instance status term` field when opening instance after updating. Fixes UIIN-3650.
 * Show "Source" field in "Version history" for a user not having permissions in Central tenant. Fixes UIIN-3686.
 * Update Inventory results list's select/unselect all checkbox label. Refs UIIN-3684.
+* Integrate Button "Generate identifier" in Inventory instance records. Refs UIIN-3679.
 
 ## [14.0.5](https://github.com/folio-org/ui-inventory/tree/v14.0.5) (2026-07-31)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v14.0.4...v14.0.5)